Use u32 instead of unsigned long since that's the type of the data that is passed to armcp15_setSingleRegionAccess().

+++ trunk/desmume/src/cp15.cpp
@@ -82,7 +82,7 @@ armcp15_t *armcp15_new(armcpu_t * c)
 #define MASKFROMREG(val)    (~((SIZEBINARY(val)-1) | 0x3F))
 #define SETFROMREG(val)     ((val) & MASKFROMREG(val))
 /* sets the precalculated regions to mask,set for the affected accesstypes */
-static void armcp15_setSingleRegionAccess(armcp15_t *armcp15,unsigned long dAccess,unsigned long iAccess,unsigned char num, unsigned long mask,unsigned long set) {
+static void armcp15_setSingleRegionAccess(armcp15_t *armcp15,u32 dAccess,u32 iAccess,unsigned char num, u32 mask,u32 set) {
 
 	switch (ACCESSTYPE(dAccess,num)) {
 		case 4: /* UNP */
